+++ CHANGES.md
@@ -7,6 +7,7 @@ See [compatibility docs](docs/compatibility.md) for details about backward compa
 * @dev
    * Enforced HTTP code for error messages fits 400-499 range
    * `validateResponseStructure()` deprecated in favor of `assertValidResponse()` 
+   * Moved Orchestra's `getPackageProviders()` out of `TestingHelpers` trait
 
 * v4.0.1 (2017-04-10)
    * TestingHelpers trait's `validateResponseStructure()` method is now public